CMS-458: Cache hit and miss count are overflowing

+++ modules/cms-core/src/main/java/com/enonic/cms/framework/cache/CacheFacade.java
@@ -66,12 +66,12 @@ public interface CacheFacade
     /**
      * Number of times a requested item was found in cache.
      */
-    public int getHitCount();
+    public long getHitCount();
 
     /**
      * Number of times a requested item was not found in cache.
      */
-    public int getMissCount();
+    public long getMissCount();
 
     /**
      * Return the number of objects in cache.
